[review] refactor: domain 패키지 하위로 review 관련 패키지 전체 이동
