+++ src/main/java/org/example/bookstoreapp/domain/review/controller/ReviewController.java
@@ -1,7 +1,7 @@
-package org.example.bookstoreapp.review.controller;
+package org.example.bookstoreapp.domain.review.controller;
 
 import lombok.RequiredArgsConstructor;
-import org.example.bookstoreapp.review.service.ReviewService;
+import org.example.bookstoreapp.domain.review.service.ReviewService;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
