+++ src/main/java/org/example/bookstoreapp/domain/review/service/ReviewService.java
@@ -1,7 +1,7 @@
-package org.example.bookstoreapp.review.service;
+package org.example.bookstoreapp.domain.review.service;
 
 import lombok.RequiredArgsConstructor;
-import org.example.bookstoreapp.review.repository.ReviewRepository;
+import org.example.bookstoreapp.domain.review.repository.ReviewRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
